requirements.txt
lightning==2.0.0
omegaconf==2.3.0
jaxtyping
typeguard
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers<0.20
transformers==4.28.1
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
wandb
gradio
git+https://github.com/ashawkey/envlight.git
torchmetrics

# deepfloyd
xformers
bitsandbytes==0.38.1
sentencepiece
safetensors
huggingface_hub

# for zero123
einops
kornia
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git

#controlnet
controlnet_aux
opencv-python-headless<4.3
rembg
pyhocon>=0.3.60
icecream>=2.1.3
